requirements.txt
openai==1.108.0
python-dotenv==1.0.1
streamlit==1.50.0

# library for images with streamlit
pillow==11.3.0
